feat: remove ApolloQueryPipe
