--- src/ApolloQueryPipe.ts
@@ -1,23 +1,0 @@
-import { Pipe } from '@angular/core';
-
-@Pipe({
-  name: 'apolloQuery',
-})
-/**
- * @deprecated since version 0.4.4 - use SelectPipe
- */
-export class ApolloQueryPipe {
-  public transform(obj: any, name: string = '') {
-    if (obj && name !== '') {
-      // for Apollo decorator
-      if (obj[name]) {
-        return obj[name];
-      }
-
-      // for Angular2Apollo.watchQuery
-      if (obj.data && obj.data[name]) {
-        return obj.data[name];
-      }
-    }
-  }
-}
