+++ src/SelectPipe.ts
@@ -1,8 +1,20 @@
-import { Pipe } from '@angular/core';
-
-import { ApolloQueryPipe } from './ApolloQueryPipe';
+import { Pipe, PipeTransform } from '@angular/core';
 
 @Pipe({
   name: 'select',
 })
-export class SelectPipe extends ApolloQueryPipe {}
+export class SelectPipe implements PipeTransform {
+  public transform(obj: any, name: string = '') {
+    if (obj && name !== '') {
+      // for Apollo decorator
+      if (obj[name]) {
+        return obj[name];
+      }
+
+      // for Angular2Apollo.watchQuery
+      if (obj.data && obj.data[name]) {
+        return obj.data[name];
+      }
+    }
+  }
+}
